requirements.txt
# Processamento de PDF
PyMuPDF==1.23.26
fitz==0.0.1.dev2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3

# Processamento de texto em português
spacy==3.7.2
nltk==3.8.1

# Utilitários
tqdm==4.66.1
colorama==0.4.6
click==8.1.7
pyyaml==6.0.1

# Desenvolvimento e testes
pytest==7.4.3
black==23.11.0
isort==5.12.0
